requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
imbalanced-learn>=0.11
smote-variants>=0.7
matplotlib>=3.7
seaborn>=0.12
scipy>=1.11
python-docx>=1.1
Pillow>=10.0
tqdm>=4.65